Fixes #85 - Int -> Using Int32 and Int64 for int/long

+++ adaptive-arp-rt/Source/Sources.Api/AccelerationListenerImpl.swift
@@ -45,7 +45,7 @@ public class AccelerationListenerImpl : BaseListenerImpl, IAccelerationListener 
 
        @param id  The id of the listener.
     */
-    public override init(id : Int) {
+    public override init(id : Int64) {
         super.init(id: id);
     }
