+++ adaptive-arp-rt/Source/Sources.Api/DatabaseResultCallbackImpl.swift
@@ -45,7 +45,7 @@ public class DatabaseResultCallbackImpl : BaseCallbackImpl, IDatabaseResultCallb
 
        @param id  The id of the callback.
     */
-    public override init(id : Int) {
+    public override init(id : Int64) {
         super.init(id: id)
     }
