+++ adaptive-arp-rt/Source/Sources.Api/DeviceOrientationListenerImpl.swift
@@ -45,7 +45,7 @@ public class DeviceOrientationListenerImpl : BaseListenerImpl, IDeviceOrientatio
 
        @param id  The id of the listener.
     */
-    public override init(id : Int) {
+    public override init(id : Int64) {
         super.init(id: id);
     }
